pyproject.toml
[tool.poetry]
name = "clout"
version = "0.1.7.post2" # PACKAGE_VERSION  (This comment helps bumpversion find the right line.)
description = "Use dataclasses without repeating yourself"
authors = ["Clout contributors <python-clout@users.noreply.github.com>"]
readme = "README.rst"
classifiers = [
        'Intended Audience :: Developers',
        'Programming Language :: Python :: 3.5',
        'Programming Language :: Python :: 3.6',
        'Programming Language :: Python :: 3.7',
        'Programming Language :: Python :: Implementation :: CPython',
        'Programming Language :: Python :: Implementation :: PyPy',
]

[tool.poetry.dependencies]
python = "^3.7"
click = "7.0"
lark-parser = "^0.7.3"
marshmallow = "^3.0"
typing-inspect = "^0.4.0"
glom = "^19.2"
inflection = "^0.3.1"
typing-extensions = "^3.7"
importlib_resources = "^1.0"
dataclasses = "^0.6.0"
appdirs = "^1.4"
attrs = "=19.2.0"




[tool.poetry.dev-dependencies]
bump2version = "^0.5.10"
coverage = "^4.5"
pytest = "^5.0"
hypothesis = "^4.32"
requests = "^2.22"
pylint = "^2.3"
twine = "^1.13"
vulture = "^1.0"
cuvner = "^18.0"
pytest-cov = "^2.7"
tox = "^3.13"
sphinx_rtd_theme = "^0.4.3"
pytest-sugar = "^0.9.2"
docshtest = "^0.0.2"
pre-commit = "^1.18"
pytest-clarity = {version = "^0.1.0-alpha.1", allows-prereleases = true}
pygments = {version = "^2.4", allows-prereleases = true}
pudb = "^2019.1"
ipython = "^7.7"
typing-inspect = "^0.4.0"
pytest-sphinx = "^0.2.2"
sphinx_autodoc_typehints = "^1.7"
marshmallow_enum = "^1.5"
marshmallow_union = "^0.1.10"
mypy = "^0.720.0"


[tool.poetry.extras]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx_autodoc_typehints",
    "sphinx-click",
    "marshmallow-jsonschema",
    "sphinx-jsonschema",

]



[tool.towncrier]

   package = "clout"
   package_dir = "src"
   filename = "CHANGELOG.rst"
   template = "changelog.d/towncrier_template.rst"
   issue_format = "`#{issue} <https://github.com/python-clout/clout/issues/{issue}>`_"
   directory = "changelog.d"
   title_format = "{version} ({project_date})"
   underlines = ["-", "^"]

   [[tool.towncrier.section]]
       path = ""

    [[tool.towncrier.type]]
        directory = "breaking"
        name = "Backward-incompatible Changes"
        showcontent = true

    [[tool.towncrier.type]]
        directory = "deprecation"
        name = "Deprecations"
        showcontent = true

    [[tool.towncrier.type]]
        directory = "change"
        name = "Changes"
        showcontent = true


[tool.black]

include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.?venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
